Resolved some requested changes.
