+++ ocean_provider/validation/requests.py
@@ -2,23 +2,21 @@
 # Copyright 2021 Ocean Protocol Foundation
 # SPDX-License-Identifier: Apache-2.0
 #
-import hashlib
-import json
+
 import os
 
 from flask import request as flask_request
 from flask_sieve import JsonRequest, ValidationException
 from flask_sieve.rules_processor import RulesProcessor
 from flask_sieve.validator import Validator
-from ocean_lib.common.agreements.service_types import ServiceTypesIndices
-from ocean_lib.web3_internal.transactions import sign_hash
 
-from ocean_provider.exceptions import InvalidSignatureError, RequestNotFound
-from ocean_provider.utils.basics import get_provider_wallet
+
+from ocean_provider.exceptions import InvalidSignatureError
+
 from ocean_provider.user_nonce import get_nonce
 from ocean_provider.utils.accounts import verify_signature
 from ocean_provider.utils.util import get_request_data
-from typing import Optional
+from ocean_provider.validation.RBACValidator import RBACValidator
 
 
 class CustomJsonRequest(JsonRequest):
@@ -76,150 +74,6 @@ class CustomValidator(Validator):
             rules, request, custom_handlers, messages, **kwargs
         )
         self._processor = CustomRulesProcessor()
-
-
-class RBACValidator:
-    def __init__(
-        self,
-        request_name=None,
-        request=None,
-        assets: Optional[list] = None,
-        algorithms: Optional[list] = None,
-    ):
-        self._request = request
-        if not request:
-            raise RequestNotFound("Request not found.")
-        action_mapping = {
-            "EncryptRequest": "encryptUrl",
-            "InitializeRequest": "initialize",
-            "DownloadRequest": "access",
-            "ComputeRequest": "compute",
-            "ComputeStartRequest": "compute",
-        }
-        self._action = action_mapping[request_name]
-        self._credentials = {
-            "type": "address",
-            "address": get_provider_wallet().address,
-        }
-        self._component = "provider"
-        self._assets = assets
-        self._algorithms = algorithms
-        if self.assets is None:
-            self._assets = []
-        if self.algorithms is None:
-            self._algorithms = []
-        self._compute_dids = {
-            "dids": [
-                {
-                    "did": asset.did,
-                    "serviceId": ServiceTypesIndices.DEFAULT_COMPUTING_INDEX,
-                }
-                for asset in self.assets
-            ],
-            "algos": [
-                {
-                    "did": algorithm.did,
-                    "serviceId": ServiceTypesIndices.DEFAULT_COMPUTING_INDEX,
-                }
-                for algorithm in self.algorithms
-            ],
-        }
-
-    @property
-    def credentials(self):
-        return self._credentials
-
-    @property
-    def provider_address(self):
-        return get_provider_wallet().address
-
-    @property
-    def component(self):
-        return self._component
-
-    @property
-    def action(self):
-        return self._action
-
-    @property
-    def assets(self):
-        return self._assets
-
-    @property
-    def request(self):
-        return self._request
-
-    @property
-    def compute_dids(self):
-        return self._compute_dids
-
-    @property
-    def algorithms(self):
-        return self._algorithms
-
-    def fails(self):
-        return False
-
-    def msg_hash(self, message: str):
-        return hashlib.sha256(message.encode("utf-8")).hexdigest()
-
-    def build_payload(self):
-        payload = {
-            "eventType": self.action,
-            "component": self.component,
-            "providerAddress": self.provider_address,
-            "credentials": self.credentials,
-        }
-        payload.update(getattr(self, f"build_{self.action}_payload")())
-        return payload
-
-    def build_encryptUrl_payload(self):
-        message = "encryptUrl" + json.dumps(self.credentials)
-        signature = sign_hash(self.msg_hash(message), get_provider_wallet())
-
-        return {"signature": signature}
-
-    def build_initialize_payload(self):
-        message = "initialize" + json.dumps(self.credentials)
-        signature = sign_hash(self.msg_hash(message), get_provider_wallet())
-        return {
-            "signature": signature,
-            "dids": [
-                {
-                    "did": asset.did,
-                    "serviceId": ServiceTypesIndices.DEFAULT_ACCESS_INDEX,
-                }
-                for asset in self.assets
-            ],
-        }
-
-    def build_access_payload(self):
-        message = "access" + json.dumps(self.credentials)
-        signature = sign_hash(self.msg_hash(message), get_provider_wallet())
-        return {
-            "signature": signature,
-            "dids": [
-                {
-                    "did": asset.did,
-                    "serviceId": ServiceTypesIndices.DEFAULT_ACCESS_INDEX,
-                }
-                for asset in self.assets
-            ],
-        }
-
-    def build_compute_payload(self):
-        message = (
-            "compute"
-            + json.dumps(self.credentials)
-            + json.dumps(self.compute_dids)
-            + json.dumps()
-        )
-        signature = sign_hash(self.msg_hash(message), get_provider_wallet())
-        return {
-            "signature": signature,
-            "dids": [compute_did for compute_did in self.compute_dids["dids"]],
-            "algos": [algo for algo in self.compute_dids["algos"]],
-        }
 
 
 class CustomRulesProcessor(RulesProcessor):
